Override creator/contributor methods to prevent Hyrax-DOI minter sending JSON
